- Fix SAKAI-833 (http://jira.hec.ca/browse/SAKAI-833) by having the ID of COConfigSerialized be managed by the DAO layer (aka Hibernate) rather than having manual ids assigned.
IDs were actually not "manually" assigned but using the Sakai IdManager class. 
This dependency has been removed and the uuid generator from Hibernate is used where relevant. The ID field length has been reduced from 99 to 32 characters to ensure that it will be compatible with a Sakai ID (which are 32-character long strings which is the hex representation of a 16 bits number). The previous code was going over the 32 characters limit, unnecessarily introducing a potential compatibility issue with Sakai. 

This also allows the OsylConfigServiceImpl class to be Saka-agnostic